feat: title and description separated from content
