+++ src/components/header/index.tsx
@@ -1,6 +1,6 @@
-import Content from './content'
 import Root from './header'
+import { Description, TextContent, Title } from './styles'
 
-const Header = { Root, Content }
+const Header = { Root, Description, TextContent, Title }
 
 export default Header
