+++ src/components/header/styles.ts
@@ -8,3 +8,22 @@ export const Container = styled.View((props) => ({
   marginHorizontal: props.theme.size.size5,
   marginVertical: props.theme.size.size6,
 }))
+
+export const TextContent = styled.View((props) => ({
+  gap: props.theme.size.size2,
+  flex: 1,
+}))
+
+export const Title = styled.Text((props) => ({
+  fontFamily: props.theme.typography.family.primary.semibold,
+  color: props.theme.palette.foreground.default,
+  fontSize: props.theme.typography.fontSize.size7,
+  lineHeight: props.theme.typography.fontSize.size8,
+}))
+
+export const Description = styled.Text((props) => ({
+  fontFamily: props.theme.typography.family.primary.regular,
+  color: props.theme.palette.foreground.default,
+  fontSize: props.theme.typography.fontSize.size4,
+  lineHeight: props.theme.typography.lineHeight.size4,
+}))
